requirements.txt
asgiref==3.5.1
Django==4.0.4
django-ckeditor==6.4.0
django-crispy-forms==1.14.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
sqlparse==0.4.2
